Fix SNES hi-res mode by not resizing gameTexture -> IOSurface in the helper.
New stuff in OEGameCoreHelper:
-bufferSize returns the maximum size of the emu video.
The IOSurface is created at this size.
-screenSize returns the current size of the active texture in the IOSurface.

The window is opened at the screenSize of the first frame. screenSize
can change but it will be resized to the original screenSize (+ filter output)
ignoring the new aspect ratio.

The QC filters (CRT+TV) don't support non-square-pixel aspect ratio, so the game
texture is still upscaled to a square pixel ratio for them, which might lead to
a slight quality loss.


git-svn-id: https://openemu.svn.sourceforge.net/svnroot/openemu/branches/iosurface@825 a16923d9-94c8-4a39-ae78-11df2d60a2ba